Refactor RssShuffleManager for spark v2 and v3 to reduce redundant code
