+++ client-spark/spark2/src/main/java/org/apache/spark/shuffle/RssShuffleManager.java
@@ -17,15 +17,10 @@
 
 package org.apache.spark.shuffle;
 
-import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.TimeUnit;
 
 import scala.Option;
 import scala.Tuple2;
@@ -34,7 +29,6 @@ import scala.collection.Seq;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.Sets;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.spark.ShuffleDependency;
 import org.apache.spark.SparkConf;
@@ -47,8 +41,6 @@ import org.apache.spark.shuffle.handle.ShuffleHandleInfo;
 import org.apache.spark.shuffle.handle.SimpleShuffleHandleInfo;
 import org.apache.spark.shuffle.handle.StageAttemptShuffleHandleInfo;
 import org.apache.spark.shuffle.reader.RssShuffleReader;
-import org.apache.spark.shuffle.writer.AddBlockEvent;
-import org.apache.spark.shuffle.writer.DataPusher;
 import org.apache.spark.shuffle.writer.RssShuffleWriter;
 import org.apache.spark.storage.BlockId;
 import org.apache.spark.storage.BlockManagerId;
@@ -56,205 +48,20 @@ import org.roaringbitmap.longlong.Roaring64NavigableMap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import org.apache.uniffle.client.impl.FailedBlockSendTracker;
+import org.apache.uniffle.client.api.ShuffleWriteClient;
 import org.apache.uniffle.client.util.ClientUtils;
 import org.apache.uniffle.common.RemoteStorageInfo;
 import org.apache.uniffle.common.ShuffleServerInfo;
-import org.apache.uniffle.common.config.RssClientConf;
-import org.apache.uniffle.common.config.RssConf;
 import org.apache.uniffle.common.exception.RssException;
 import org.apache.uniffle.common.exception.RssFetchFailedException;
-import org.apache.uniffle.common.rpc.GrpcServer;
-import org.apache.uniffle.common.util.BlockIdLayout;
-import org.apache.uniffle.common.util.JavaUtils;
-import org.apache.uniffle.common.util.RssUtils;
-import org.apache.uniffle.common.util.ThreadUtils;
 import org.apache.uniffle.shuffle.RssShuffleClientFactory;
 import org.apache.uniffle.shuffle.manager.RssShuffleManagerBase;
-import org.apache.uniffle.shuffle.manager.ShuffleManagerGrpcService;
-import org.apache.uniffle.shuffle.manager.ShuffleManagerServerFactory;
-
-import static org.apache.spark.shuffle.RssSparkConfig.RSS_BLOCK_ID_SELF_MANAGEMENT_ENABLED;
-import static org.apache.spark.shuffle.RssSparkConfig.RSS_RESUBMIT_STAGE_WITH_FETCH_FAILURE_ENABLED;
-import static org.apache.spark.shuffle.RssSparkConfig.RSS_RESUBMIT_STAGE_WITH_WRITE_FAILURE_ENABLED;
-import static org.apache.uniffle.common.config.RssBaseConf.RPC_SERVER_PORT;
-import static org.apache.uniffle.common.config.RssClientConf.MAX_CONCURRENCY_PER_PARTITION_TO_WRITE;
 
 public class RssShuffleManager extends RssShuffleManagerBase {
-
   private static final Logger LOG = LoggerFactory.getLogger(RssShuffleManager.class);
-  private final long heartbeatInterval;
-  private final long heartbeatTimeout;
-  private ScheduledExecutorService heartBeatScheduledExecutorService;
-  private Map<String, Set<Long>> taskToSuccessBlockIds = JavaUtils.newConcurrentMap();
-  private Map<String, FailedBlockSendTracker> taskToFailedBlockSendTracker =
-      JavaUtils.newConcurrentMap();
-  private final int dataReplica;
-  private final int dataReplicaWrite;
-  private final int dataReplicaRead;
-  private final boolean dataReplicaSkipEnabled;
-  private final int dataTransferPoolSize;
-  private final int dataCommitPoolSize;
-  private Set<String> failedTaskIds = Sets.newConcurrentHashSet();
-  private boolean heartbeatStarted = false;
-  private final int maxFailures;
-  private final boolean speculation;
-  private final BlockIdLayout blockIdLayout;
-  private final String user;
-  private final String uuid;
-  private DataPusher dataPusher;
-  private GrpcServer shuffleManagerServer;
-  private ShuffleManagerGrpcService service;
 
   public RssShuffleManager(SparkConf sparkConf, boolean isDriver) {
-    if (sparkConf.getBoolean("spark.sql.adaptive.enabled", false)) {
-      throw new IllegalArgumentException(
-          "Spark2 doesn't support AQE, spark.sql.adaptive.enabled should be false.");
-    }
-    this.sparkConf = sparkConf;
-    this.user = sparkConf.get("spark.rss.quota.user", "user");
-    this.uuid = sparkConf.get("spark.rss.quota.uuid", Long.toString(System.currentTimeMillis()));
-    this.dynamicConfEnabled = sparkConf.get(RssSparkConfig.RSS_DYNAMIC_CLIENT_CONF_ENABLED);
-
-    // fetch client conf and apply them if necessary
-    if (isDriver && this.dynamicConfEnabled) {
-      fetchAndApplyDynamicConf(sparkConf);
-    }
-    RssSparkShuffleUtils.validateRssClientConf(sparkConf);
-
-    // configure block id layout
-    this.maxFailures = sparkConf.getInt("spark.task.maxFailures", 4);
-    this.speculation = sparkConf.getBoolean("spark.speculation", false);
-    RssConf rssConf = RssSparkConfig.toRssConf(sparkConf);
-    RssUtils.setExtraJavaProperties(rssConf);
-    // configureBlockIdLayout requires maxFailures and speculation to be initialized
-    configureBlockIdLayout(sparkConf, rssConf);
-    this.blockIdLayout = BlockIdLayout.from(rssConf);
-
-    // set & check replica config
-    this.dataReplica = sparkConf.get(RssSparkConfig.RSS_DATA_REPLICA);
-    this.dataReplicaWrite = sparkConf.get(RssSparkConfig.RSS_DATA_REPLICA_WRITE);
-    this.dataReplicaRead = sparkConf.get(RssSparkConfig.RSS_DATA_REPLICA_READ);
-    this.dataTransferPoolSize = sparkConf.get(RssSparkConfig.RSS_DATA_TRANSFER_POOL_SIZE);
-    this.dataReplicaSkipEnabled = sparkConf.get(RssSparkConfig.RSS_DATA_REPLICA_SKIP_ENABLED);
-    this.maxConcurrencyPerPartitionToWrite =
-        RssSparkConfig.toRssConf(sparkConf).get(MAX_CONCURRENCY_PER_PARTITION_TO_WRITE);
-    LOG.info(
-        "Check quorum config [{}:{}:{}:{}]",
-        dataReplica,
-        dataReplicaWrite,
-        dataReplicaRead,
-        dataReplicaSkipEnabled);
-    RssUtils.checkQuorumSetting(dataReplica, dataReplicaWrite, dataReplicaRead);
-
-    this.clientType = sparkConf.get(RssSparkConfig.RSS_CLIENT_TYPE);
-    this.heartbeatInterval = sparkConf.get(RssSparkConfig.RSS_HEARTBEAT_INTERVAL);
-    this.heartbeatTimeout =
-        sparkConf.getLong(RssSparkConfig.RSS_HEARTBEAT_TIMEOUT.key(), heartbeatInterval / 2);
-    int retryMax = sparkConf.get(RssSparkConfig.RSS_CLIENT_RETRY_MAX);
-    long retryIntervalMax = sparkConf.get(RssSparkConfig.RSS_CLIENT_RETRY_INTERVAL_MAX);
-    int heartBeatThreadNum = sparkConf.get(RssSparkConfig.RSS_CLIENT_HEARTBEAT_THREAD_NUM);
-    this.dataCommitPoolSize = sparkConf.get(RssSparkConfig.RSS_DATA_COMMIT_POOL_SIZE);
-    int unregisterThreadPoolSize =
-        sparkConf.get(RssSparkConfig.RSS_CLIENT_UNREGISTER_THREAD_POOL_SIZE);
-    int unregisterTimeoutSec = sparkConf.get(RssSparkConfig.RSS_CLIENT_UNREGISTER_TIMEOUT_SEC);
-    int unregisterRequestTimeoutSec =
-        sparkConf.get(RssSparkConfig.RSS_CLIENT_UNREGISTER_REQUEST_TIMEOUT_SEC);
-    // External shuffle service is not supported when using remote shuffle service
-    sparkConf.set("spark.shuffle.service.enabled", "false");
-    LOG.info("Disable external shuffle service in RssShuffleManager.");
-    // If we store shuffle data in distributed filesystem or in a disaggregated
-    // shuffle cluster, we don't need shuffle data locality
-    sparkConf.set("spark.shuffle.reduceLocality.enabled", "false");
-    LOG.info("Disable shuffle data locality in RssShuffleManager.");
-
-    this.shuffleIdToPartitionNum = JavaUtils.newConcurrentMap();
-    this.shuffleIdToNumMapTasks = JavaUtils.newConcurrentMap();
-    // stage retry for write/fetch failure
-    rssStageRetryForFetchFailureEnabled =
-        rssConf.get(RSS_RESUBMIT_STAGE_WITH_FETCH_FAILURE_ENABLED);
-    rssStageRetryForWriteFailureEnabled =
-        rssConf.get(RSS_RESUBMIT_STAGE_WITH_WRITE_FAILURE_ENABLED);
-    if (rssStageRetryForFetchFailureEnabled || rssStageRetryForWriteFailureEnabled) {
-      rssStageRetryEnabled = true;
-      List<String> logTips = new ArrayList<>();
-      if (rssStageRetryForWriteFailureEnabled) {
-        logTips.add("write");
-      }
-      if (rssStageRetryForWriteFailureEnabled) {
-        logTips.add("fetch");
-      }
-      LOG.info(
-          "Activate the stage retry mechanism that will resubmit stage on {} failure",
-          StringUtils.join(logTips, "/"));
-    }
-    this.partitionReassignEnabled = rssConf.getBoolean(RssClientConf.RSS_CLIENT_REASSIGN_ENABLED);
-    this.blockIdSelfManagedEnabled = rssConf.getBoolean(RSS_BLOCK_ID_SELF_MANAGEMENT_ENABLED);
-    this.shuffleManagerRpcServiceEnabled =
-        partitionReassignEnabled || rssStageRetryEnabled || blockIdSelfManagedEnabled;
-    if (!sparkConf.getBoolean(RssSparkConfig.RSS_TEST_FLAG.key(), false)) {
-      if (isDriver) {
-        heartBeatScheduledExecutorService =
-            ThreadUtils.getDaemonSingleThreadScheduledExecutor("rss-heartbeat");
-        if (shuffleManagerRpcServiceEnabled) {
-          LOG.info("stage resubmit is supported and enabled");
-          // start shuffle manager server
-          rssConf.set(RPC_SERVER_PORT, 0);
-          ShuffleManagerServerFactory factory = new ShuffleManagerServerFactory(this, rssConf);
-          service = factory.getService();
-          shuffleManagerServer = factory.getServer(service);
-          try {
-            shuffleManagerServer.start();
-            // pass this as a spark.rss.shuffle.manager.grpc.port config, so it can be propagated to
-            // executor properly.
-            sparkConf.set(
-                RssSparkConfig.RSS_SHUFFLE_MANAGER_GRPC_PORT, shuffleManagerServer.getPort());
-          } catch (Exception e) {
-            LOG.error("Failed to start shuffle manager server", e);
-            throw new RssException(e);
-          }
-        }
-      }
-      if (shuffleManagerRpcServiceEnabled) {
-        getOrCreateShuffleManagerClientSupplier();
-      }
-      this.shuffleWriteClient =
-          RssShuffleClientFactory.getInstance()
-              .createShuffleWriteClient(
-                  RssShuffleClientFactory.newWriteBuilder()
-                      .blockIdSelfManagedEnabled(blockIdSelfManagedEnabled)
-                      .managerClientSupplier(managerClientSupplier)
-                      .clientType(clientType)
-                      .retryMax(retryMax)
-                      .retryIntervalMax(retryIntervalMax)
-                      .heartBeatThreadNum(heartBeatThreadNum)
-                      .replica(dataReplica)
-                      .replicaWrite(dataReplicaWrite)
-                      .replicaRead(dataReplicaRead)
-                      .replicaSkipEnabled(dataReplicaSkipEnabled)
-                      .dataTransferPoolSize(dataTransferPoolSize)
-                      .dataCommitPoolSize(dataCommitPoolSize)
-                      .unregisterThreadPoolSize(unregisterThreadPoolSize)
-                      .unregisterTimeSec(unregisterTimeoutSec)
-                      .unregisterRequestTimeSec(unregisterRequestTimeoutSec)
-                      .rssConf(rssConf));
-      registerCoordinator();
-
-      // for non-driver executor, start a thread for sending shuffle data to shuffle server
-      LOG.info("RSS data pusher is starting...");
-      int poolSize = sparkConf.get(RssSparkConfig.RSS_CLIENT_SEND_THREAD_POOL_SIZE);
-      int keepAliveTime = sparkConf.get(RssSparkConfig.RSS_CLIENT_SEND_THREAD_POOL_KEEPALIVE);
-      this.dataPusher =
-          new DataPusher(
-              shuffleWriteClient,
-              taskToSuccessBlockIds,
-              taskToFailedBlockSendTracker,
-              failedTaskIds,
-              poolSize,
-              keepAliveTime);
-    }
-    this.shuffleHandleInfoManager = new ShuffleHandleInfoManager();
-    this.rssStageResubmitManager = new RssStageResubmitManager();
+    super(sparkConf, isDriver);
   }
 
   // This method is called in Spark driver side,
@@ -380,42 +187,6 @@ public class RssShuffleManager extends RssShuffleManagerBase {
     return new RssShuffleHandle(shuffleId, appId, numMaps, dependency, hdlInfoBd);
   }
 
-  private void startHeartbeat() {
-    shuffleWriteClient.registerApplicationInfo(appId, heartbeatTimeout, user);
-    if (!sparkConf.getBoolean(RssSparkConfig.RSS_TEST_FLAG.key(), false) && !heartbeatStarted) {
-      heartBeatScheduledExecutorService.scheduleAtFixedRate(
-          () -> {
-            try {
-              shuffleWriteClient.sendAppHeartbeat(appId, heartbeatTimeout);
-              LOG.info("Finish send heartbeat to coordinator and servers");
-            } catch (Exception e) {
-              LOG.warn("Fail to send heartbeat to coordinator and servers", e);
-            }
-          },
-          heartbeatInterval / 2,
-          heartbeatInterval,
-          TimeUnit.MILLISECONDS);
-      heartbeatStarted = true;
-    }
-  }
-
-  @VisibleForTesting
-  protected void registerCoordinator() {
-    String coordinators = sparkConf.get(RssSparkConfig.RSS_COORDINATOR_QUORUM.key());
-    LOG.info("Registering coordinators {}", coordinators);
-    shuffleWriteClient.registerCoordinators(
-        coordinators,
-        this.sparkConf.get(RssSparkConfig.RSS_CLIENT_RETRY_INTERVAL_MAX),
-        this.sparkConf.get(RssSparkConfig.RSS_CLIENT_RETRY_MAX));
-  }
-
-  public CompletableFuture<Long> sendData(AddBlockEvent event) {
-    if (dataPusher != null && event != null) {
-      return dataPusher.send(event);
-    }
-    return new CompletableFuture<>();
-  }
-
   // This method is called in Spark executor,
   // getting information from Spark driver via the ShuffleHandle.
   @Override
@@ -445,24 +216,6 @@ public class RssShuffleManager extends RssShuffleManagerBase {
     } else {
       throw new RssException("Unexpected ShuffleHandle:" + handle.getClass().getName());
     }
-  }
-
-  /**
-   * Derives block id layout config from maximum number of allowed partitions. Computes the number
-   * of required bits for partition id and task attempt id and reserves remaining bits for sequence
-   * number.
-   *
-   * @param sparkConf Spark config providing max partitions
-   * @param rssConf Rss config to amend
-   */
-  public void configureBlockIdLayout(SparkConf sparkConf, RssConf rssConf) {
-    configureBlockIdLayout(sparkConf, rssConf, maxFailures, speculation);
-  }
-
-  @Override
-  public long getTaskAttemptIdForBlockId(int mapIndex, int attemptNo) {
-    return getTaskAttemptIdForBlockId(
-        mapIndex, attemptNo, maxFailures, speculation, blockIdLayout.taskAttemptIdBits);
   }
 
   // This method is called in Spark executor,
@@ -564,44 +317,6 @@ public class RssShuffleManager extends RssShuffleManagerBase {
   }
 
   @Override
-  public boolean unregisterShuffle(int shuffleId) {
-    try {
-      super.unregisterShuffle(shuffleId);
-      if (SparkEnv.get().executorId().equals("driver")) {
-        shuffleWriteClient.unregisterShuffle(appId, shuffleId);
-        shuffleIdToNumMapTasks.remove(shuffleId);
-        shuffleIdToPartitionNum.remove(shuffleId);
-        if (service != null) {
-          service.unregisterShuffle(shuffleId);
-        }
-      }
-    } catch (Exception e) {
-      LOG.warn("Errors on unregistering from remote shuffle-servers", e);
-    }
-    return true;
-  }
-
-  @Override
-  public void stop() {
-    super.stop();
-    if (heartBeatScheduledExecutorService != null) {
-      heartBeatScheduledExecutorService.shutdownNow();
-    }
-    if (dataPusher != null) {
-      try {
-        dataPusher.close();
-      } catch (IOException e) {
-        LOG.warn("Errors on closing data pusher", e);
-      }
-    }
-    if (shuffleWriteClient != null) {
-      // Unregister shuffle before closing shuffle write client.
-      shuffleWriteClient.unregisterShuffle(appId);
-      shuffleWriteClient.close();
-    }
-  }
-
-  @Override
   public ShuffleBlockResolver shuffleBlockResolver() {
     throw new RssException("RssShuffleManager.shuffleBlockResolver is not implemented");
   }
@@ -631,91 +346,15 @@ public class RssShuffleManager extends RssShuffleManagerBase {
     return taskIdBitmap;
   }
 
-  public Set<Long> getFailedBlockIds(String taskId) {
-    FailedBlockSendTracker blockIdsFailedSendTracker = getBlockIdsFailedSendTracker(taskId);
-    if (blockIdsFailedSendTracker == null) {
-      return Collections.emptySet();
-    }
-    return blockIdsFailedSendTracker.getFailedBlockIds();
-  }
-
-  public Set<Long> getSuccessBlockIds(String taskId) {
-    Set<Long> result = taskToSuccessBlockIds.get(taskId);
-    if (result == null) {
-      result = Collections.emptySet();
-    }
-    return result;
-  }
-
-  @VisibleForTesting
-  public void addSuccessBlockIds(String taskId, Set<Long> blockIds) {
-    if (taskToSuccessBlockIds.get(taskId) == null) {
-      taskToSuccessBlockIds.put(taskId, Sets.newHashSet());
-    }
-    taskToSuccessBlockIds.get(taskId).addAll(blockIds);
-  }
-
-  @VisibleForTesting
-  public void addFailedBlockSendTracker(
-      String taskId, FailedBlockSendTracker failedBlockSendTracker) {
-    taskToFailedBlockSendTracker.putIfAbsent(taskId, failedBlockSendTracker);
-  }
-
-  public void clearTaskMeta(String taskId) {
-    taskToSuccessBlockIds.remove(taskId);
-    taskToFailedBlockSendTracker.remove(taskId);
-  }
-
-  @VisibleForTesting
-  public SparkConf getSparkConf() {
-    return sparkConf;
-  }
-
   @VisibleForTesting
   public void setAppId(String appId) {
     this.appId = appId;
-  }
-
-  public boolean markFailedTask(String taskId) {
-    LOG.info("Mark the task: {} failed.", taskId);
-    failedTaskIds.add(taskId);
-    return true;
-  }
-
-  public boolean isValidTask(String taskId) {
-    return !failedTaskIds.contains(taskId);
-  }
-
-  public DataPusher getDataPusher() {
-    return dataPusher;
-  }
-
-  public void setDataPusher(DataPusher dataPusher) {
-    this.dataPusher = dataPusher;
   }
 
   /** @return the unique spark id for rss shuffle */
   @Override
   public String getAppId() {
     return appId;
-  }
-
-  /**
-   * @param shuffleId the shuffleId to query
-   * @return the num of partitions(a.k.a reduce tasks) for shuffle with shuffle id.
-   */
-  @Override
-  public int getPartitionNum(int shuffleId) {
-    return shuffleIdToPartitionNum.getOrDefault(shuffleId, 0);
-  }
-
-  /**
-   * @param shuffleId the shuffle id to query
-   * @return the num of map tasks for current shuffle with shuffle id.
-   */
-  @Override
-  public int getNumMaps(int shuffleId) {
-    return shuffleIdToNumMapTasks.getOrDefault(shuffleId, 0);
   }
 
   private Roaring64NavigableMap getShuffleResult(
@@ -740,10 +379,6 @@ public class RssShuffleManager extends RssShuffleManagerBase {
     }
   }
 
-  public FailedBlockSendTracker getBlockIdsFailedSendTracker(String taskId) {
-    return taskToFailedBlockSendTracker.get(taskId);
-  }
-
   private ShuffleServerInfo assignShuffleServer(int shuffleId, String faultyShuffleServerId) {
     Set<String> faultyServerIds = Sets.newHashSet(faultyShuffleServerId);
     faultyServerIds.addAll(rssStageResubmitManager.getServerIdBlackList());
@@ -753,5 +388,45 @@ public class RssShuffleManager extends RssShuffleManagerBase {
       return partitionToServers.get(0).get(0);
     }
     return null;
+  }
+
+  @Override
+  protected ShuffleWriteClient createShuffleWriteClient() {
+    int unregisterThreadPoolSize =
+        sparkConf.get(RssSparkConfig.RSS_CLIENT_UNREGISTER_THREAD_POOL_SIZE);
+    int unregisterTimeoutSec = sparkConf.get(RssSparkConfig.RSS_CLIENT_UNREGISTER_TIMEOUT_SEC);
+    int unregisterRequestTimeoutSec =
+        sparkConf.get(RssSparkConfig.RSS_CLIENT_UNREGISTER_REQUEST_TIMEOUT_SEC);
+    long retryIntervalMax = sparkConf.get(RssSparkConfig.RSS_CLIENT_RETRY_INTERVAL_MAX);
+    int heartBeatThreadNum = sparkConf.get(RssSparkConfig.RSS_CLIENT_HEARTBEAT_THREAD_NUM);
+
+    final int retryMax = sparkConf.get(RssSparkConfig.RSS_CLIENT_RETRY_MAX);
+    return RssShuffleClientFactory.getInstance()
+        .createShuffleWriteClient(
+            RssShuffleClientFactory.newWriteBuilder()
+                .blockIdSelfManagedEnabled(blockIdSelfManagedEnabled)
+                .managerClientSupplier(managerClientSupplier)
+                .clientType(clientType)
+                .retryMax(retryMax)
+                .retryIntervalMax(retryIntervalMax)
+                .heartBeatThreadNum(heartBeatThreadNum)
+                .replica(dataReplica)
+                .replicaWrite(dataReplicaWrite)
+                .replicaRead(dataReplicaRead)
+                .replicaSkipEnabled(dataReplicaSkipEnabled)
+                .dataTransferPoolSize(dataTransferPoolSize)
+                .dataCommitPoolSize(dataCommitPoolSize)
+                .unregisterThreadPoolSize(unregisterThreadPoolSize)
+                .unregisterTimeSec(unregisterTimeoutSec)
+                .unregisterRequestTimeSec(unregisterRequestTimeoutSec)
+                .rssConf(rssConf));
+  }
+
+  @Override
+  protected void checkSupported(SparkConf sparkConf) {
+    if (sparkConf.getBoolean("spark.sql.adaptive.enabled", false)) {
+      throw new IllegalArgumentException(
+          "Spark2 doesn't support AQE, spark.sql.adaptive.enabled should be false.");
+    }
   }
 }
