+++ client-spark/spark3/src/main/java/org/apache/spark/shuffle/RssShuffleManager.java
@@ -17,16 +17,10 @@
 
 package org.apache.spark.shuffle;
 
-import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 import java.util.Set;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Collectors;
 
@@ -38,7 +32,6 @@ import scala.collection.Seq;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.Sets;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-import org.apache.commons.lang3.StringUtils;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.spark.MapOutputTracker;
 import org.apache.spark.ShuffleDependency;
@@ -53,7 +46,6 @@ import org.apache.spark.shuffle.handle.ShuffleHandleInfo;
 import org.apache.spark.shuffle.handle.SimpleShuffleHandleInfo;
 import org.apache.spark.shuffle.handle.StageAttemptShuffleHandleInfo;
 import org.apache.spark.shuffle.reader.RssShuffleReader;
-import org.apache.spark.shuffle.writer.AddBlockEvent;
 import org.apache.spark.shuffle.writer.DataPusher;
 import org.apache.spark.shuffle.writer.RssShuffleWriter;
 import org.apache.spark.sql.internal.SQLConf;
@@ -64,6 +56,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import org.apache.uniffle.client.PartitionDataReplicaRequirementTracking;
+import org.apache.uniffle.client.api.ShuffleWriteClient;
 import org.apache.uniffle.client.impl.FailedBlockSendTracker;
 import org.apache.uniffle.client.util.ClientUtils;
 import org.apache.uniffle.common.RemoteStorageInfo;
@@ -73,235 +66,16 @@ import org.apache.uniffle.common.config.RssClientConf;
 import org.apache.uniffle.common.config.RssConf;
 import org.apache.uniffle.common.exception.RssException;
 import org.apache.uniffle.common.exception.RssFetchFailedException;
-import org.apache.uniffle.common.rpc.GrpcServer;
-import org.apache.uniffle.common.util.BlockIdLayout;
-import org.apache.uniffle.common.util.JavaUtils;
 import org.apache.uniffle.common.util.RssUtils;
-import org.apache.uniffle.common.util.ThreadUtils;
 import org.apache.uniffle.shuffle.RssShuffleClientFactory;
 import org.apache.uniffle.shuffle.manager.RssShuffleManagerBase;
-import org.apache.uniffle.shuffle.manager.ShuffleManagerGrpcService;
-import org.apache.uniffle.shuffle.manager.ShuffleManagerServerFactory;
-
-import static org.apache.spark.shuffle.RssSparkConfig.RSS_BLOCK_ID_SELF_MANAGEMENT_ENABLED;
-import static org.apache.spark.shuffle.RssSparkConfig.RSS_PARTITION_REASSIGN_MAX_REASSIGNMENT_SERVER_NUM;
-import static org.apache.spark.shuffle.RssSparkConfig.RSS_RESUBMIT_STAGE_WITH_FETCH_FAILURE_ENABLED;
-import static org.apache.spark.shuffle.RssSparkConfig.RSS_RESUBMIT_STAGE_WITH_WRITE_FAILURE_ENABLED;
-import static org.apache.uniffle.common.config.RssBaseConf.RPC_SERVER_PORT;
-import static org.apache.uniffle.common.config.RssClientConf.MAX_CONCURRENCY_PER_PARTITION_TO_WRITE;
 
 public class RssShuffleManager extends RssShuffleManagerBase {
   private static final Logger LOG = LoggerFactory.getLogger(RssShuffleManager.class);
-  private final long heartbeatInterval;
-  private final long heartbeatTimeout;
-  private final int dataReplica;
-  private final int dataReplicaWrite;
-  private final int dataReplicaRead;
-  private final boolean dataReplicaSkipEnabled;
-  private final int dataTransferPoolSize;
-  private final int dataCommitPoolSize;
-  private final Map<String, Set<Long>> taskToSuccessBlockIds;
-  private final Map<String, FailedBlockSendTracker> taskToFailedBlockSendTracker;
-  private ScheduledExecutorService heartBeatScheduledExecutorService;
-  private boolean heartbeatStarted = false;
-  private final BlockIdLayout blockIdLayout;
-  private final int maxFailures;
-  private final boolean speculation;
-  private String user;
-  private String uuid;
-  private Set<String> failedTaskIds = Sets.newConcurrentHashSet();
-  private DataPusher dataPusher;
-  private ShuffleManagerGrpcService service;
-  private GrpcServer shuffleManagerServer;
 
   public RssShuffleManager(SparkConf conf, boolean isDriver) {
-    this.sparkConf = conf;
-    boolean supportsRelocation =
-        Optional.ofNullable(SparkEnv.get())
-            .map(env -> env.serializer().supportsRelocationOfSerializedObjects())
-            .orElse(true);
-    if (!supportsRelocation) {
-      LOG.warn(
-          "RSSShuffleManager requires a serializer which supports relocations of serialized object. Please set "
-              + "spark.serializer to org.apache.spark.serializer.KryoSerializer instead");
-    }
-    this.user = sparkConf.get("spark.rss.quota.user", "user");
-    this.uuid = sparkConf.get("spark.rss.quota.uuid", Long.toString(System.currentTimeMillis()));
-    this.dynamicConfEnabled = sparkConf.get(RssSparkConfig.RSS_DYNAMIC_CLIENT_CONF_ENABLED);
-
-    // fetch client conf and apply them if necessary
-    if (isDriver && this.dynamicConfEnabled) {
-      fetchAndApplyDynamicConf(sparkConf);
-    }
-    RssSparkShuffleUtils.validateRssClientConf(sparkConf);
-
-    // set & check replica config
-    this.dataReplica = sparkConf.get(RssSparkConfig.RSS_DATA_REPLICA);
-    this.dataReplicaWrite = sparkConf.get(RssSparkConfig.RSS_DATA_REPLICA_WRITE);
-    this.dataReplicaRead = sparkConf.get(RssSparkConfig.RSS_DATA_REPLICA_READ);
-    this.dataReplicaSkipEnabled = sparkConf.get(RssSparkConfig.RSS_DATA_REPLICA_SKIP_ENABLED);
-    LOG.info(
-        "Check quorum config ["
-            + dataReplica
-            + ":"
-            + dataReplicaWrite
-            + ":"
-            + dataReplicaRead
-            + ":"
-            + dataReplicaSkipEnabled
-            + "]");
-    RssUtils.checkQuorumSetting(dataReplica, dataReplicaWrite, dataReplicaRead);
-
-    this.heartbeatInterval = sparkConf.get(RssSparkConfig.RSS_HEARTBEAT_INTERVAL);
-    this.heartbeatTimeout =
-        sparkConf.getLong(RssSparkConfig.RSS_HEARTBEAT_TIMEOUT.key(), heartbeatInterval / 2);
-    final int retryMax = sparkConf.get(RssSparkConfig.RSS_CLIENT_RETRY_MAX);
-    this.clientType = sparkConf.get(RssSparkConfig.RSS_CLIENT_TYPE);
+    super(conf, isDriver);
     this.dataDistributionType = getDataDistributionType(sparkConf);
-    RssConf rssConf = RssSparkConfig.toRssConf(sparkConf);
-    RssUtils.setExtraJavaProperties(rssConf);
-    this.maxConcurrencyPerPartitionToWrite = rssConf.get(MAX_CONCURRENCY_PER_PARTITION_TO_WRITE);
-    this.maxFailures = sparkConf.getInt("spark.task.maxFailures", 4);
-    this.speculation = sparkConf.getBoolean("spark.speculation", false);
-    // configureBlockIdLayout requires maxFailures and speculation to be initialized
-    configureBlockIdLayout(sparkConf, rssConf);
-    this.blockIdLayout = BlockIdLayout.from(rssConf);
-    this.dataTransferPoolSize = sparkConf.get(RssSparkConfig.RSS_DATA_TRANSFER_POOL_SIZE);
-    this.dataCommitPoolSize = sparkConf.get(RssSparkConfig.RSS_DATA_COMMIT_POOL_SIZE);
-    // External shuffle service is not supported when using remote shuffle service
-    sparkConf.set("spark.shuffle.service.enabled", "false");
-    sparkConf.set("spark.dynamicAllocation.shuffleTracking.enabled", "false");
-    sparkConf.set(RssSparkConfig.RSS_ENABLED.key(), "true");
-    LOG.info("Disable external shuffle service in RssShuffleManager.");
-    sparkConf.set("spark.sql.adaptive.localShuffleReader.enabled", "false");
-    LOG.info("Disable local shuffle reader in RssShuffleManager.");
-    // If we store shuffle data in distributed filesystem or in a disaggregated
-    // shuffle cluster, we don't need shuffle data locality
-    sparkConf.set("spark.shuffle.reduceLocality.enabled", "false");
-    LOG.info("Disable shuffle data locality in RssShuffleManager.");
-    taskToSuccessBlockIds = JavaUtils.newConcurrentMap();
-    taskToFailedBlockSendTracker = JavaUtils.newConcurrentMap();
-    this.shuffleIdToPartitionNum = JavaUtils.newConcurrentMap();
-    this.shuffleIdToNumMapTasks = JavaUtils.newConcurrentMap();
-    this.partitionReassignEnabled = rssConf.get(RssClientConf.RSS_CLIENT_REASSIGN_ENABLED);
-
-    // stage retry for write/fetch failure
-    rssStageRetryForFetchFailureEnabled =
-        rssConf.get(RSS_RESUBMIT_STAGE_WITH_FETCH_FAILURE_ENABLED);
-    rssStageRetryForWriteFailureEnabled =
-        rssConf.get(RSS_RESUBMIT_STAGE_WITH_WRITE_FAILURE_ENABLED);
-    if (rssStageRetryForFetchFailureEnabled || rssStageRetryForWriteFailureEnabled) {
-      rssStageRetryEnabled = true;
-      List<String> logTips = new ArrayList<>();
-      if (rssStageRetryForWriteFailureEnabled) {
-        logTips.add("write");
-      }
-      if (rssStageRetryForWriteFailureEnabled) {
-        logTips.add("fetch");
-      }
-      LOG.info(
-          "Activate the stage retry mechanism that will resubmit stage on {} failure",
-          StringUtils.join(logTips, "/"));
-    }
-
-    // The feature of partition reassign is exclusive with multiple replicas and stage retry.
-    if (partitionReassignEnabled) {
-      if (dataReplica > 1) {
-        throw new RssException(
-            "The feature of task partition reassign is incompatible with multiple replicas mechanism.");
-      }
-    }
-
-    this.blockIdSelfManagedEnabled = rssConf.getBoolean(RSS_BLOCK_ID_SELF_MANAGEMENT_ENABLED);
-    this.shuffleManagerRpcServiceEnabled =
-        partitionReassignEnabled || rssStageRetryEnabled || blockIdSelfManagedEnabled;
-    if (isDriver) {
-      heartBeatScheduledExecutorService =
-          ThreadUtils.getDaemonSingleThreadScheduledExecutor("rss-heartbeat");
-      if (shuffleManagerRpcServiceEnabled) {
-        LOG.info("stage resubmit is supported and enabled");
-        // start shuffle manager server
-        rssConf.set(RPC_SERVER_PORT, 0);
-        ShuffleManagerServerFactory factory = new ShuffleManagerServerFactory(this, rssConf);
-        service = factory.getService();
-        shuffleManagerServer = factory.getServer(service);
-        try {
-          shuffleManagerServer.start();
-          // pass this as a spark.rss.shuffle.manager.grpc.port config, so it can be propagated to
-          // executor properly.
-          sparkConf.set(
-              RssSparkConfig.RSS_SHUFFLE_MANAGER_GRPC_PORT, shuffleManagerServer.getPort());
-        } catch (Exception e) {
-          LOG.error("Failed to start shuffle manager server", e);
-          throw new RssException(e);
-        }
-      }
-    }
-    if (shuffleManagerRpcServiceEnabled) {
-      getOrCreateShuffleManagerClientSupplier();
-    }
-    int unregisterThreadPoolSize =
-        sparkConf.get(RssSparkConfig.RSS_CLIENT_UNREGISTER_THREAD_POOL_SIZE);
-    int unregisterTimeoutSec = sparkConf.get(RssSparkConfig.RSS_CLIENT_UNREGISTER_TIMEOUT_SEC);
-    int unregisterRequestTimeoutSec =
-        sparkConf.get(RssSparkConfig.RSS_CLIENT_UNREGISTER_REQUEST_TIMEOUT_SEC);
-    long retryIntervalMax = sparkConf.get(RssSparkConfig.RSS_CLIENT_RETRY_INTERVAL_MAX);
-    int heartBeatThreadNum = sparkConf.get(RssSparkConfig.RSS_CLIENT_HEARTBEAT_THREAD_NUM);
-    shuffleWriteClient =
-        RssShuffleClientFactory.getInstance()
-            .createShuffleWriteClient(
-                RssShuffleClientFactory.newWriteBuilder()
-                    .blockIdSelfManagedEnabled(blockIdSelfManagedEnabled)
-                    .managerClientSupplier(managerClientSupplier)
-                    .clientType(clientType)
-                    .retryMax(retryMax)
-                    .retryIntervalMax(retryIntervalMax)
-                    .heartBeatThreadNum(heartBeatThreadNum)
-                    .replica(dataReplica)
-                    .replicaWrite(dataReplicaWrite)
-                    .replicaRead(dataReplicaRead)
-                    .replicaSkipEnabled(dataReplicaSkipEnabled)
-                    .dataTransferPoolSize(dataTransferPoolSize)
-                    .dataCommitPoolSize(dataCommitPoolSize)
-                    .unregisterThreadPoolSize(unregisterThreadPoolSize)
-                    .unregisterTimeSec(unregisterTimeoutSec)
-                    .unregisterRequestTimeSec(unregisterRequestTimeoutSec)
-                    .rssConf(rssConf));
-    registerCoordinator();
-
-    LOG.info("Rss data pusher is starting...");
-    int poolSize = sparkConf.get(RssSparkConfig.RSS_CLIENT_SEND_THREAD_POOL_SIZE);
-    int keepAliveTime = sparkConf.get(RssSparkConfig.RSS_CLIENT_SEND_THREAD_POOL_KEEPALIVE);
-    this.dataPusher =
-        new DataPusher(
-            shuffleWriteClient,
-            taskToSuccessBlockIds,
-            taskToFailedBlockSendTracker,
-            failedTaskIds,
-            poolSize,
-            keepAliveTime);
-    this.partitionReassignMaxServerNum =
-        rssConf.get(RSS_PARTITION_REASSIGN_MAX_REASSIGNMENT_SERVER_NUM);
-    this.shuffleHandleInfoManager = new ShuffleHandleInfoManager();
-    this.rssStageResubmitManager = new RssStageResubmitManager();
-  }
-
-  public CompletableFuture<Long> sendData(AddBlockEvent event) {
-    if (dataPusher != null && event != null) {
-      return dataPusher.send(event);
-    }
-    return new CompletableFuture<>();
-  }
-
-  @VisibleForTesting
-  protected static ShuffleDataDistributionType getDataDistributionType(SparkConf sparkConf) {
-    RssConf rssConf = RssSparkConfig.toRssConf(sparkConf);
-    if ((boolean) sparkConf.get(SQLConf.ADAPTIVE_EXECUTION_ENABLED())
-        && !rssConf.containsKey(RssClientConf.DATA_DISTRIBUTION_TYPE.key())) {
-      return ShuffleDataDistributionType.LOCAL_ORDER;
-    }
-
-    return rssConf.get(RssClientConf.DATA_DISTRIBUTION_TYPE);
   }
 
   // For testing only
@@ -312,72 +86,7 @@ public class RssShuffleManager extends RssShuffleManagerBase {
       DataPusher dataPusher,
       Map<String, Set<Long>> taskToSuccessBlockIds,
       Map<String, FailedBlockSendTracker> taskToFailedBlockSendTracker) {
-    this.sparkConf = conf;
-    this.clientType = sparkConf.get(RssSparkConfig.RSS_CLIENT_TYPE);
-    RssConf rssConf = RssSparkConfig.toRssConf(sparkConf);
-    this.dataDistributionType = rssConf.get(RssClientConf.DATA_DISTRIBUTION_TYPE);
-    this.blockIdLayout = BlockIdLayout.from(rssConf);
-    this.maxConcurrencyPerPartitionToWrite = rssConf.get(MAX_CONCURRENCY_PER_PARTITION_TO_WRITE);
-    this.maxFailures = sparkConf.getInt("spark.task.maxFailures", 4);
-    this.speculation = sparkConf.getBoolean("spark.speculation", false);
-    // configureBlockIdLayout requires maxFailures and speculation to be initialized
-    configureBlockIdLayout(sparkConf, rssConf);
-    this.heartbeatInterval = sparkConf.get(RssSparkConfig.RSS_HEARTBEAT_INTERVAL);
-    this.heartbeatTimeout =
-        sparkConf.getLong(RssSparkConfig.RSS_HEARTBEAT_TIMEOUT.key(), heartbeatInterval / 2);
-    this.dataReplica = sparkConf.get(RssSparkConfig.RSS_DATA_REPLICA);
-    this.dataReplicaWrite = sparkConf.get(RssSparkConfig.RSS_DATA_REPLICA_WRITE);
-    this.dataReplicaRead = sparkConf.get(RssSparkConfig.RSS_DATA_REPLICA_READ);
-    this.dataReplicaSkipEnabled = sparkConf.get(RssSparkConfig.RSS_DATA_REPLICA_SKIP_ENABLED);
-    LOG.info(
-        "Check quorum config ["
-            + dataReplica
-            + ":"
-            + dataReplicaWrite
-            + ":"
-            + dataReplicaRead
-            + ":"
-            + dataReplicaSkipEnabled
-            + "]");
-    RssUtils.checkQuorumSetting(dataReplica, dataReplicaWrite, dataReplicaRead);
-
-    int retryMax = sparkConf.get(RssSparkConfig.RSS_CLIENT_RETRY_MAX);
-    long retryIntervalMax = sparkConf.get(RssSparkConfig.RSS_CLIENT_RETRY_INTERVAL_MAX);
-    int heartBeatThreadNum = sparkConf.get(RssSparkConfig.RSS_CLIENT_HEARTBEAT_THREAD_NUM);
-    this.dataTransferPoolSize = sparkConf.get(RssSparkConfig.RSS_DATA_TRANSFER_POOL_SIZE);
-    this.dataCommitPoolSize = sparkConf.get(RssSparkConfig.RSS_DATA_COMMIT_POOL_SIZE);
-    int unregisterThreadPoolSize =
-        sparkConf.get(RssSparkConfig.RSS_CLIENT_UNREGISTER_THREAD_POOL_SIZE);
-    int unregisterTimeoutSec = sparkConf.get(RssSparkConfig.RSS_CLIENT_UNREGISTER_TIMEOUT_SEC);
-    int unregisterRequestTimeoutSec =
-        sparkConf.get(RssSparkConfig.RSS_CLIENT_UNREGISTER_REQUEST_TIMEOUT_SEC);
-    shuffleWriteClient =
-        RssShuffleClientFactory.getInstance()
-            .createShuffleWriteClient(
-                RssShuffleClientFactory.getInstance()
-                    .newWriteBuilder()
-                    .clientType(clientType)
-                    .retryMax(retryMax)
-                    .retryIntervalMax(retryIntervalMax)
-                    .heartBeatThreadNum(heartBeatThreadNum)
-                    .replica(dataReplica)
-                    .replicaWrite(dataReplicaWrite)
-                    .replicaRead(dataReplicaRead)
-                    .replicaSkipEnabled(dataReplicaSkipEnabled)
-                    .dataTransferPoolSize(dataTransferPoolSize)
-                    .dataCommitPoolSize(dataCommitPoolSize)
-                    .unregisterThreadPoolSize(unregisterThreadPoolSize)
-                    .unregisterTimeSec(unregisterTimeoutSec)
-                    .unregisterRequestTimeSec(unregisterRequestTimeoutSec)
-                    .rssConf(rssConf));
-    this.taskToSuccessBlockIds = taskToSuccessBlockIds;
-    this.heartBeatScheduledExecutorService = null;
-    this.taskToFailedBlockSendTracker = taskToFailedBlockSendTracker;
-    this.dataPusher = dataPusher;
-    this.partitionReassignMaxServerNum =
-        rssConf.get(RSS_PARTITION_REASSIGN_MAX_REASSIGNMENT_SERVER_NUM);
-    this.shuffleHandleInfoManager = new ShuffleHandleInfoManager();
-    this.rssStageResubmitManager = new RssStageResubmitManager();
+    super(conf, isDriver, dataPusher, taskToSuccessBlockIds, taskToFailedBlockSendTracker);
   }
 
   // This method is called in Spark driver side,
@@ -525,17 +234,6 @@ public class RssShuffleManager extends RssShuffleManagerBase {
         rssHandle,
         this::markFailedTask,
         context);
-  }
-
-  @Override
-  public void configureBlockIdLayout(SparkConf sparkConf, RssConf rssConf) {
-    configureBlockIdLayout(sparkConf, rssConf, maxFailures, speculation);
-  }
-
-  @Override
-  public long getTaskAttemptIdForBlockId(int mapIndex, int attemptNo) {
-    return getTaskAttemptIdForBlockId(
-        mapIndex, attemptNo, maxFailures, speculation, blockIdLayout.taskAttemptIdBits);
   }
 
   public void setPusherAppId(RssShuffleHandle rssShuffleHandle) {
@@ -846,126 +544,51 @@ public class RssShuffleManager extends RssShuffleManagerBase {
   }
 
   @Override
-  public boolean unregisterShuffle(int shuffleId) {
-    try {
-      super.unregisterShuffle(shuffleId);
-      if (SparkEnv.get().executorId().equals("driver")) {
-        shuffleWriteClient.unregisterShuffle(id.get(), shuffleId);
-        shuffleIdToPartitionNum.remove(shuffleId);
-        shuffleIdToNumMapTasks.remove(shuffleId);
-        if (service != null) {
-          service.unregisterShuffle(shuffleId);
-        }
-      }
-    } catch (Exception e) {
-      LOG.warn("Errors on unregistering from remote shuffle-servers", e);
-    }
-    return true;
-  }
-
-  @Override
   public ShuffleBlockResolver shuffleBlockResolver() {
     throw new RssException("RssShuffleManager.shuffleBlockResolver is not implemented");
   }
 
   @Override
-  public void stop() {
-    super.stop();
-    if (heartBeatScheduledExecutorService != null) {
-      heartBeatScheduledExecutorService.shutdownNow();
-    }
-    if (shuffleWriteClient != null) {
-      // Unregister shuffle before closing shuffle write client.
-      shuffleWriteClient.unregisterShuffle(getAppId());
-      shuffleWriteClient.close();
-    }
-    if (dataPusher != null) {
-      try {
-        dataPusher.close();
-      } catch (IOException e) {
-        LOG.warn("Errors on closing data pusher", e);
-      }
-    }
+  protected ShuffleWriteClient createShuffleWriteClient() {
+    int unregisterThreadPoolSize =
+        sparkConf.get(RssSparkConfig.RSS_CLIENT_UNREGISTER_THREAD_POOL_SIZE);
+    int unregisterTimeoutSec = sparkConf.get(RssSparkConfig.RSS_CLIENT_UNREGISTER_TIMEOUT_SEC);
+    int unregisterRequestTimeoutSec =
+        sparkConf.get(RssSparkConfig.RSS_CLIENT_UNREGISTER_REQUEST_TIMEOUT_SEC);
+    long retryIntervalMax = sparkConf.get(RssSparkConfig.RSS_CLIENT_RETRY_INTERVAL_MAX);
+    int heartBeatThreadNum = sparkConf.get(RssSparkConfig.RSS_CLIENT_HEARTBEAT_THREAD_NUM);
 
-    if (shuffleManagerServer != null) {
-      try {
-        shuffleManagerServer.stop();
-      } catch (InterruptedException e) {
-        // ignore
-        LOG.info("shuffle manager server is interrupted during stop");
-      }
-    }
-  }
-
-  public void clearTaskMeta(String taskId) {
-    taskToSuccessBlockIds.remove(taskId);
-    taskToFailedBlockSendTracker.remove(taskId);
+    final int retryMax = sparkConf.get(RssSparkConfig.RSS_CLIENT_RETRY_MAX);
+    return RssShuffleClientFactory.getInstance()
+        .createShuffleWriteClient(
+            RssShuffleClientFactory.newWriteBuilder()
+                .blockIdSelfManagedEnabled(blockIdSelfManagedEnabled)
+                .managerClientSupplier(managerClientSupplier)
+                .clientType(clientType)
+                .retryMax(retryMax)
+                .retryIntervalMax(retryIntervalMax)
+                .heartBeatThreadNum(heartBeatThreadNum)
+                .replica(dataReplica)
+                .replicaWrite(dataReplicaWrite)
+                .replicaRead(dataReplicaRead)
+                .replicaSkipEnabled(dataReplicaSkipEnabled)
+                .dataTransferPoolSize(dataTransferPoolSize)
+                .dataCommitPoolSize(dataCommitPoolSize)
+                .unregisterThreadPoolSize(unregisterThreadPoolSize)
+                .unregisterTimeSec(unregisterTimeoutSec)
+                .unregisterRequestTimeSec(unregisterRequestTimeoutSec)
+                .rssConf(rssConf));
   }
 
   @VisibleForTesting
-  protected void registerCoordinator() {
-    String coordinators = sparkConf.get(RssSparkConfig.RSS_COORDINATOR_QUORUM.key());
-    LOG.info("Start Registering coordinators {}", coordinators);
-    shuffleWriteClient.registerCoordinators(
-        coordinators,
-        this.sparkConf.get(RssSparkConfig.RSS_CLIENT_RETRY_INTERVAL_MAX),
-        this.sparkConf.get(RssSparkConfig.RSS_CLIENT_RETRY_MAX));
-  }
-
-  private synchronized void startHeartbeat() {
-    shuffleWriteClient.registerApplicationInfo(id.get(), heartbeatTimeout, user);
-    if (!heartbeatStarted) {
-      heartBeatScheduledExecutorService.scheduleAtFixedRate(
-          () -> {
-            try {
-              String appId = id.get();
-              shuffleWriteClient.sendAppHeartbeat(appId, heartbeatTimeout);
-              LOG.info("Finish send heartbeat to coordinator and servers");
-            } catch (Exception e) {
-              LOG.warn("Fail to send heartbeat to coordinator and servers", e);
-            }
-          },
-          heartbeatInterval / 2,
-          heartbeatInterval,
-          TimeUnit.MILLISECONDS);
-      heartbeatStarted = true;
+  protected static ShuffleDataDistributionType getDataDistributionType(SparkConf sparkConf) {
+    RssConf rssConf = RssSparkConfig.toRssConf(sparkConf);
+    if ((boolean) sparkConf.get(SQLConf.ADAPTIVE_EXECUTION_ENABLED())
+        && !rssConf.containsKey(RssClientConf.DATA_DISTRIBUTION_TYPE.key())) {
+      return ShuffleDataDistributionType.LOCAL_ORDER;
     }
-  }
 
-  public Set<Long> getFailedBlockIds(String taskId) {
-    FailedBlockSendTracker blockIdsFailedSendTracker = getBlockIdsFailedSendTracker(taskId);
-    if (blockIdsFailedSendTracker == null) {
-      return Collections.emptySet();
-    }
-    return blockIdsFailedSendTracker.getFailedBlockIds();
-  }
-
-  public Set<Long> getSuccessBlockIds(String taskId) {
-    Set<Long> result = taskToSuccessBlockIds.get(taskId);
-    if (result == null) {
-      result = Collections.emptySet();
-    }
-    return result;
-  }
-
-  /** @return the unique spark id for rss shuffle */
-  @Override
-  public String getAppId() {
-    return id.get();
-  }
-
-  @Override
-  public int getPartitionNum(int shuffleId) {
-    return shuffleIdToPartitionNum.getOrDefault(shuffleId, 0);
-  }
-
-  /**
-   * @param shuffleId the shuffle id to query
-   * @return the num of map tasks for current shuffle with shuffle id.
-   */
-  @Override
-  public int getNumMaps(int shuffleId) {
-    return shuffleIdToNumMapTasks.getOrDefault(shuffleId, 0);
+    return rssConf.get(RssClientConf.DATA_DISTRIBUTION_TYPE);
   }
 
   static class ReadMetrics extends ShuffleReadMetrics {
@@ -1019,16 +642,6 @@ public class RssShuffleManager extends RssShuffleManagerBase {
     this.id = new AtomicReference<>(appId);
   }
 
-  public boolean markFailedTask(String taskId) {
-    LOG.info("Mark the task: {} failed.", taskId);
-    failedTaskIds.add(taskId);
-    return true;
-  }
-
-  public boolean isValidTask(String taskId) {
-    return !failedTaskIds.contains(taskId);
-  }
-
   private Roaring64NavigableMap getShuffleResultForMultiPart(
       String clientType,
       Map<ShuffleServerInfo, Set<Integer>> serverToPartitions,
@@ -1049,24 +662,5 @@ public class RssShuffleManager extends RssShuffleManagerBase {
       throw RssSparkShuffleUtils.reportRssFetchFailedException(
           managerClientSupplier, e, sparkConf, appId, shuffleId, stageAttemptId, failedPartitions);
     }
-  }
-
-  public FailedBlockSendTracker getBlockIdsFailedSendTracker(String taskId) {
-    return taskToFailedBlockSendTracker.get(taskId);
-  }
-
-  @VisibleForTesting
-  public void setDataPusher(DataPusher dataPusher) {
-    this.dataPusher = dataPusher;
-  }
-
-  @VisibleForTesting
-  public Map<String, Set<Long>> getTaskToSuccessBlockIds() {
-    return taskToSuccessBlockIds;
-  }
-
-  @VisibleForTesting
-  public Map<String, FailedBlockSendTracker> getTaskToFailedBlockSendTracker() {
-    return taskToFailedBlockSendTracker;
   }
 }
